requirements.txt
numpy==1.21.3
numpy-quaternion==2022.4.1
